requirements.txt
pydantic>=2.0.0
pika>=1.3.0
python-dotenv==1.2.1
pytest==8.3.4
testcontainers>=4.13.3
